Test/setup

+++ packages/staking/contracts/libraries/Errors.sol
@@ -15,9 +15,9 @@ library Errors {
         uint256 contractBalance,
         uint256 requiredAmount
     );
-    error Staking__InvalidDistributionPercentage(uint24 percentage);
-    error Staking__BandAllocationExceedsMaximum(uint24 percentage);
-    error Staking__TotalAllocationExceedsMaximum(uint24 percentage);
+    error Staking__InvalidDistributionPercentage(uint48 percentage);
+    error Staking__BandAllocationExceedsMaximum(uint48 percentage);
+    error Staking__TotalAllocationExceedsMaximum(uint48 percentage);
     error Staking__InvalidBandsAmount();
     error Staking__ZeroAmount();
     error Staking__ZeroAddress();
